+++ profiles/serializers/activity_serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from ..models import Activity
+
+
+class ActivitySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Activity
+        fields = "__all__"
