+++ profiles/views/regions_views.py
@@ -4,29 +4,14 @@ from rest_framework.generics import (
 )
 from rest_framework.permissions import IsAdminUser
 from ..permissions import RequestIsReadOnly
-from ..models import Activity, Region
-from ..serializers.serializers import (
-    ActivitySerializer,
-    RegionSerializer,
-)
-
-
-class ActivityList(ListCreateAPIView):
-    serializer_class = ActivitySerializer
-    permission_classes = (RequestIsReadOnly | IsAdminUser,)
-    queryset = Activity.objects.all()
+from ..models import Region
+from ..serializers.region_serializers import RegionSerializer
 
 
 class RegionList(ListCreateAPIView):
     serializer_class = RegionSerializer
     permission_classes = (RequestIsReadOnly | IsAdminUser,)
     queryset = Region.objects.all()
-
-
-class ActivityDetail(RetrieveUpdateDestroyAPIView):
-    serializer_class = ActivitySerializer
-    permission_classes = (IsAdminUser,)
-    queryset = Activity.objects.all()
 
 
 class RegionDetail(RetrieveUpdateDestroyAPIView):
